PointModelResult can be interpolated
